Split firestore.js to each class files
