+++ CHANGELOGS.md
@@ -1,0 +1,54 @@
+# History
+
+## 1.3.1
+- Organize file structure.
+
+## 1.3.0
+- Merge multi-project config changes
+- Merge sub document changes
+- Update Web SDK reference to 5.2.0
+- Introduce QueryDataSnapshot
+- Implement timestampsInSnapshots option in configuration
+
+## 1.2.0
+- Update Android dependency versions
+- Update iOS dependency versions
+- Update plugin dependencies
+- WARNING: The Android update may require you to update com.google.gms:google-services to 4.0.0, com.android.tools.build:gradle to 3.1.2 and gradle to 4.4.4 (look in platforms/android/cordova/lib/builders/GradleBuilder.js)
+
+## 1.1.0
+- Add support for FieldValue
+- Add experimental support for Transactions. _Please note this is **experimental**!_
+- Add startswith polyfill
+
+## 1.0.10
+- Correct log level when creating results
+
+## 1.0.9
+- Updated Dependencies
+- Remove incorrect Java 7 dependency
+
+## 1.0.8
+- Ensure dates work for queries and nested data
+- Implement delete()
+- Update README
+
+## 1.0.7
+- Remove dependency on cordova-plugin-firebase-hooks
+
+## 1.0.6
+- Correct README History
+- Make browser firebase dependency loading smarter
+
+## 1.0.5
+## 1.0.4
+## 1.0.3
+- Address plugin dependency issues
+
+## 1.0.2
+- Updated version
+- Added firebase hooks dependency
+- Corrected iOS source/header-file config
+
+## 1.0.0
+Initial release
