+++ README.md
@@ -14,36 +14,100 @@ This plugin supports the following platforms:
 - Browser
 
 # Installation
-`cordova plugin add cordova-plugin-firestore --variable ANDROID_FIREBASE_CORE_VERSION=16.0.0 --variable ANDROID_FIREBASE_FIRESTORE_VERSION=17.0.1 --save`
 
-or
+  - **Step1** Install this plugin
 
-`phonegap plugin add cordova-plugin-firestore --variable ANDROID_FIREBASE_CORE_VERSION=16.0.0 --variable ANDROID_FIREBASE_FIRESTORE_VERSION=17.0.1`
+  ```bash
+  $> cordova plugin add cordova-plugin-firestore --save
+  ```
 
-Omitting `FIREBASE_VERSION` will use a default value.
+  - **Step2** Download `google-services.json`, then put it at `(your_project_dir)/google-services.json`
+    Hint: [Get a config file for your Android App](https://support.google.com/firebase/answer/7015592#android)
 
-## Firebase configuration
-### Android
-You must ensure that `google-services.json` is put in the correct location. This can be achieved using the following in your `config.xml`:
+    Then add the below three lines into `(your_project_dir)/config.xml` file.
 
-```
-<platform name="android">
-    <resource-file src="google-services.json" target="google-services.json" />
-</platform>
-```
+    ```xml
+    <platform name="android">
+        <resource-file src="google-services.json" target="app/google-services.json" />
+    </platform>
+    ```
 
-#### Dependencies
-##### cordova-support-google-services
+  - **Step3** Download `GoogleService-Info.plist`, then put it at `(your_project_dir)/GoogleService-Info.plist`
+    Hint: [Get a config file for your iOS App](https://support.google.com/firebase/answer/7015592#ios)
 
-In order to ensure Firebase initialises correctly on Android this plugin can be used. This is not automatically added as a dependency to allow for the configuration it performs to be done manually if desired.
+    Then add the below three lines into `(your_project_dir)/config.xml` file.
 
-### iOS
-iOS requires `GoogleService-Info.plist` is put in the correct location. Similarly this can be done as follows:
-```
-<platform name="ios">
-    <resource-file src="GoogleService-Info.plist" />
-</platform>
-```
+    ```xml
+    <platform name="ios">
+        <resource-file src="GoogleService-Info.plist" />
+    </platform>
+    ```
+
+  - **Step4** Make sure your Firestore rule
+
+    ```
+    service cloud.firestore {
+      match /databases/{database}/documents {
+        match /{document=**} {
+          allow read, write : if request.auth != null;  // This is an example.
+        }
+      }
+    }
+    ```
+
+  - **Step5** Write HelloWorld code
+    ```js
+    var options = {
+      "datePrefix": '__DATE:',
+      "fieldValueDelete": "__DELETE",
+      "fieldValueServerTimestamp" : "__SERVERTIMESTAMP",
+      "persist": true,
+      "config" : {}
+    };
+
+    if (cordova.platformId === "browser") {
+
+      options.config = {
+        apiKey: "AIzaSyB1Tqdt15UXdQTuvgD1_KkbSX0xTaUtpcw",
+        authDomain: "localhost",
+        projectId: "geofire-example-572f9"
+      };
+    }
+
+    Firestore.initialise(options).then(function(db) {
+      // Add a second document with a generated ID.
+      db.collection("users").add({
+          first: "Alan",
+          middle: "Mathison",
+          last: "Turing",
+          born: 1912
+      })
+      .then(function(docRef) {
+          console.log("Document written with ID: ", docRef.id);
+      })
+      .catch(function(error) {
+          console.error("Error adding document: ", error);
+      });
+    });
+    ```
+
+  - **Step6** Let's run it!
+    ```
+    $> cordova platform add browser // android, ios
+
+    $> cordova run browser
+    ```
+
+## Install optional variables
+
+  - `Android` **ANDROID_FIREBASE_CORE_VERSION = (16.0.3)**<br>
+    The `com.google.firebase:firebase-core` version.
+    You can find the latest version at [here](https://firebase.google.com/docs/android/setup#available_libraries).
+
+  - `Android` **ANDROID_FIREBASE_FIRESTORE_VERSION = (17.1.0)**<br>
+    The `com.google.firebase:firebase-firestore` version.
+    You can find the latest version at [here](https://firebase.google.com/docs/android/setup#available_libraries).
+
 #### Keychain Sharing Capability
 If using multiple Firebase plugins it may be necessary to enable this.
 
@@ -155,7 +219,7 @@ The client will receive the field as a Javascript Date.
 This conversion also happens when specify a field in a where condition.
 
 ### timestampsInSnapshots
-This option is provided for compatability with upcoming Firestore changes. By default this option is set to `false` but you
+This option is provided for compatibility with upcoming Firestore changes. By default this option is set to `false` but you
 are advised to heed the following if you wish for your app to continue to work in the browser implementation:
 
 ```
@@ -196,54 +260,3 @@ I have learnt a number of things whilst implementing this:
 - The documentation states that the database cannot be initialised in a seperate thread when using persistence. In my experience this should say it cannot be *used* in multiple threads.
 - When used on Android ensure that at least `com.google.gms:google-services:3.1.1` is used in build dependencies. Earlier versions did not work for me.
 - Yes, I did spell initialise() with an 's' - I am from the UK
-
-# History
-## 1.3.0
-- Merge multi-project config changes
-- Merge sub document changes
-- Update Web SDK reference to 5.2.0
-- Introduce QueryDataSnapshot
-- Implement timestampsInSnapshots option in configuration
-
-## 1.2.0
-- Update Android dependency versions
-- Update iOS dependency versions
-- Update plugin dependencies
-- WARNING: The Android update may require you to update com.google.gms:google-services to 4.0.0, com.android.tools.build:gradle to 3.1.2 and gradle to 4.4.4 (look in platforms/android/cordova/lib/builders/GradleBuilder.js)
-
-## 1.1.0
-- Add support for FieldValue
-- Add experimental support for Transactions. _Please note this is **experimental**!_
-- Add startswith polyfill
-
-## 1.0.10
-- Correct log level when creating results
-
-## 1.0.9
-- Updated Dependencies
-- Remove incorrect Java 7 dependency
-
-## 1.0.8
-- Ensure dates work for queries and nested data
-- Implement delete()
-- Update README
-
-## 1.0.7
-- Remove dependency on cordova-plugin-firebase-hooks
-
-## 1.0.6
-- Correct README History
-- Make browser firebase dependency loading smarter
-
-## 1.0.5
-## 1.0.4
-## 1.0.3
-- Address plugin dependency issues
-
-## 1.0.2
-- Updated version
-- Added firebase hooks dependency
-- Corrected iOS source/header-file config
-
-## 1.0.0
-Initial release
