+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cordova-plugin-firestore",
-  "version": "1.3.0",
+  "version": "1.3.1",
   "cordova": {
     "id": "cordova-plugin-firestore",
     "platforms": [
