+++ www/android_ios/firestore_options.js
@@ -1,0 +1,9 @@
+var FirestoreOptions = {
+  "datePrefix": "__DATE:",
+  "fieldValueDelete": "__DELETE",
+  "fieldValueServerTimestamp": "__SERVERTIMESTAMP",
+  "persist": true,
+  "timestampsInSnapshots": false
+};
+
+module.exports = FirestoreOptions;
