+++ www/android_ios/firestore_timestamp.js
@@ -1,0 +1,11 @@
+function FirestoreTimestamp(timestamp) {
+  this._timestamp = timestamp;
+}
+
+FirestoreTimestamp.prototype = {
+  toDate: function() {
+    return new Date(this._timestamp);
+  }
+};
+
+module.exports = FirestoreTimestamp;
