+++ www/android_ios/query_snapshot.js
@@ -1,0 +1,49 @@
+var QueryDocumentSnapshot = require("./QueryDocumentSnapshot");
+
+function QuerySnapshot(data) {
+  this._data = data;
+}
+
+QuerySnapshot.prototype = {
+  forEach: function (callback, thisArg) {
+    var keys = Object.keys(this._data.docs);
+    for (var i = 0; i < keys.length; i++) {
+      callback(new QueryDocumentSnapshot(this._data.docs[i]));
+    }
+  }
+};
+
+Object.defineProperties(QuerySnapshot.prototype, {
+  docChanges: {
+    get: function () {
+      throw "QuerySnapshot.docChanges: Not supported";
+    }
+  },
+  docs: {
+    get: function () {
+      return this._data.docs;
+    }
+  },
+  empty: {
+    get: function () {
+      return this._data.docs.length === 0;
+    }
+  },
+  metadata: {
+    get: function () {
+      throw "QuerySnapshot.metadata: Not supported";
+    }
+  },
+  query: {
+    get: function () {
+      throw "QuerySnapshot.query: Not supported";
+    }
+  },
+  size: {
+    get: function () {
+      return this._data.docs.length;
+    }
+  }
+});
+
+module.exports = QuerySnapshot;
